requirements.txt
asgiref==3.6.0
boto3==1.26.69
botocore==1.29.69
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.6
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-on-heroku==1.1.2
django-storages==1.13.2
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-yasg==1.21.5
environs==9.5.0
graphviz==0.20.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pydot==1.4.2
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0